refactor: Refatora o ReadersController para incluir o ResponseEntity nas rotas e customizar a resposta das requisicoes do cliente

+++ src/main/java/dev/cesar/LibrarySystem/Readers/ReadersController.java
@@ -1,5 +1,7 @@
 package dev.cesar.LibrarySystem.Readers;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -20,28 +22,49 @@ public class ReadersController {
     }
 
     @PostMapping("/create")
-    public ReadersDTO createReader(@RequestBody ReadersDTO reader) {
-        return readersService.createReader(reader);
+    public ResponseEntity<String> createReader(@RequestBody ReadersDTO reader) {
+        ReadersDTO newReader = readersService.createReader(reader);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body("Reader successfully created: " + newReader.getName() + ". ID: " + newReader.getId());
     }
 
     @GetMapping("/list")
-    public List<ReadersDTO> listAllReaders() {
-        return readersService.listAllReaders();
+    public ResponseEntity<List<ReadersDTO>> listAllReaders() {
+        List<ReadersDTO> readers = readersService.listAllReaders();
+        return ResponseEntity.ok(readers);
     }
 
     @GetMapping("/list/{id}")
-    public ReadersDTO listReadersById(@PathVariable Long id) {
-        return readersService.listReadersById(id);
+    public ResponseEntity<?> listReadersById(@PathVariable Long id) {
+        ReadersDTO reader = readersService.listReadersById(id);
+        if (reader != null) {
+            return ResponseEntity.status(HttpStatus.FOUND)
+                    .body(reader);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Reader of ID: " + id + " not found.");
+        }
     }
 
     @PutMapping("/update/{id}")
-    public ReadersDTO updateReaderById(@PathVariable Long id, @RequestBody ReadersDTO reader) {
-        return readersService.updateReaderById(id, reader);
+    public ResponseEntity<String> updateReaderById(@PathVariable Long id, @RequestBody ReadersDTO reader) {
+        ReadersDTO updatedReader = readersService.updateReaderById(id, reader);
+        if (updatedReader != null) {
+            return ResponseEntity.ok("Reader of ID: " + id + " updated.");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Reader of ID: " + id + " not found.");
+        }
     }
 
     @DeleteMapping("/delete/{id}")
-    public void deleteReaderById(@PathVariable Long id) {
-        readersService.deleteReaderById(id);
+    public ResponseEntity<String> deleteReaderById(@PathVariable Long id) {
+        if (readersService.listReadersById(id) != null) {
+            readersService.deleteReaderById(id);
+            return ResponseEntity.ok("Reader of ID: " + id + " deleted.");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Reader of ID: " + id + " not found.");
+        }
     }
-
 }
